requirements.txt
beautifulsoup4==4.5.1
lxml==4.9.2
numpy==1.24.2
pandas==2.0.0
playsound==1.3.0
py-vt==0.0.6
pyobjc-core==10.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.12.4
simpleaudio==1.0.4
six==1.16.0
tzdata==2023.3
vt-timetable==0.2.6
